Fix documentation in the SSL part.
Fix to the documentation on the SSL part. …
trustAll parameter can be set alone (so the keyStore related attribute
are not required). Additionally need to be set as a string otherwise
there is an error and the config file is not taken into consideration.
Error happens also when defining the attribute as boolean in the feature
file.