Upgrade snakeyaml from 1.33 to 2.0 fixing CVE-2022-1471
Snakeyaml 2.0 has a fix for this Arbitrary Code Execution vulnerability:
https://nvd.nist.gov/vuln/detail/CVE-2022-1471

Karate is not affected by this vulnerability, but other code might be
so that Karate should be compatible with Snakeyaml 2.0 and by default
ships with Snakeyaml 2.0.

The JsonUtils code change is required for 2.0 but also works in 1.33
so that projects that use latest Karate may manually downgrade Snakeyaml
from 2.0 to 1.33 if needed.